// src/interfaces/redux/action-types.ts
export enum actionTypes {
    FAILURE = "FAILURE",
    INCREMENT = "INCREMENT",
    DECREMENT = "DECREMENT",
    RESET = "RESET",

    START_CLOCK = "START_CLOCK",
    TICK_CLOCK = "TICK_CLOCK",

    LOAD_DATA = "LOAD_DATA",
    LOAD_DATA_SUCCESS = "LOAD_DATA_SUCCESS",

    LOAD_DATA_COMPLEX = "LOAD_DATA_COMPLEX",
    LOAD_DATA_COMPLEX_SUCCESS = "LOAD_DATA_COMPLEX_SUCCESS",

    POST_EXAMPLE = "POST_EXAMPLE",
    POST_EXAMPLE_SUCCESS = "POST_EXAMPLE_SUCCESS"
}